pyproject.toml
[tool.poetry]
name = "sawyer-bot"
version = "0.1.0"
description = ""
authors = ["Voronwe Sul <mail@voronwe.me>"]
packages = [{include = "sawyer_bot"}]

[tool.poetry.dependencies]
python =  ">=3.8,<3.12"
python-telegram-bot = { version = "^20.2", extras = ["webhooks"] }
scipy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
